Added new Widget module and updated styles
A new Widget module has been created and added to the app.ts file aimed to handle a new feature in our application which was not present before. Adjustments have also been made in our style.css file. The change in .btn:hover and .btn:active classes aims to enhance UI/UX by modifying the box-shadow property which makes the button more distinguishable when it is hovered over or clicked. Also, notice that a new .myty class has been added, this was to provide styling for the new elements that our Widget module might require.